Spark 3.2: Simplify action class hierarchy
