--- spark/v3.2/spark/src/main/java/org/apache/iceberg/spark/actions/BaseRewriteDataFilesSpark3Action.java
@@ -1,48 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.iceberg.spark.actions;
-
-import org.apache.iceberg.Table;
-import org.apache.iceberg.actions.BinPackStrategy;
-import org.apache.iceberg.actions.RewriteDataFiles;
-import org.apache.iceberg.actions.SortStrategy;
-import org.apache.spark.sql.SparkSession;
-
-public class BaseRewriteDataFilesSpark3Action extends BaseRewriteDataFilesSparkAction {
-
-  protected BaseRewriteDataFilesSpark3Action(SparkSession spark, Table table) {
-    super(spark, table);
-  }
-
-  @Override
-  protected BinPackStrategy binPackStrategy() {
-    return new Spark3BinPackStrategy(table(), spark());
-  }
-
-  @Override
-  protected SortStrategy sortStrategy() {
-    return new Spark3SortStrategy(table(), spark());
-  }
-
-  @Override
-  protected RewriteDataFiles self() {
-    return this;
-  }
-}
